pyproject.toml
[tool.poetry]
name = "MusicLedStudio"
version = "2.0.94"
description = "A network based light effect controller"
authors = ["MusicLedStudio Developers"]
license = "GPL-3.0 License"
readme = "README.rst"
packages = [
	{include = "mls"},
	{include = "mls_frontend"},
    {include = "mls_assets"},
    {include = "loopback"},
]
include = [
    "README.rst",
    "AUTHORS.rst",
    "CHANGELOG.rst",
    "LICENSE.txt",
    "mls.env"
]

[tool.poetry.scripts]
mls = "mls.__main__:main"
mls-loopback-install = "loopback.__main__:copy_lib"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
numpy = "^1.26.2"
cffi = ">=0.16.0"
wheel = ">=0.36.2"
aiohttp = "^3.9.1"
aiohttp-cors = ">=0.7.0"
aubio = ">=0.4.9"
cython = ">=3.0.7"
certifi = ">=2023.11.17"
multidict = ">=6.0.4"
openrgb-python = ">=0.2.15"
paho-mqtt = ">=1.6.1"
psutil = ">=5.9.7"
pyserial = ">=3.5"
pystray = ">=0.19.5"
python-rtmidi = ">=1.5.8"
requests = ">=2.31.0"
sacn = ">=1.9.0"
sentry-sdk = ">=1.40.4"
sounddevice = ">=0.4.6"
samplerate = ">=0.2.1"
icmplib = ">=3.0.4"
voluptuous = ">=0.14.1"
zeroconf = ">=0.131.0"
pillow = ">=10.1.0"
flux-led = ">=1.0.4"
python-osc = ">=1.8.3"
pybase64 = "~=1.3.1"
mss = "~=9.0.1"
setuptools = "~=69.1.0"
uvloop = {version = ">=0.16.0", markers = "sys_platform != 'win32'"}
rpi-ws281x = {version = ">=4.3.0", platform = "linux"}
python-mbedtls = {version = "^2.8.0", markers = "(sys_platform == 'linux' and platform_machine != 'aarch64' and platform_machine != 'armv7l') or sys_platform == 'win32' or sys_platform == 'darwin'", optional = true}
stupidartnet = "^1.4.0"
python-dotenv = "^1.0.0"


[tool.poetry.extras]
hue = ["python-mbedtls"]

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
flake8 = ">=7.0.0,<7.1.0"
black = ">=24.3.0,<24.4.0"
isort = ">=5.13.2,<5.14.0"
pre-commit = ">=3.6.0,<3.7.0"
bokeh = "^3.3.3"
pytest = "^8.0.0"
pytest-order = "^1.2.0"
pyinstaller = "^6.3.0"


[tool.pytest.ini_options]
addopts = "tests"
testpaths = "tests"
norecursedirs = [
    "dist",
    "build",
    ".tox",
]
filterwarnings = "ignore::DeprecationWarning"


[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79
# The following are the default values applied when profile="black"
# multi_line_output = 3
# include_trailing_comma = True
# force_grid_wrap = 0
# use_parentheses = True
# ensure_newline_before_comments = True

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
